agregue el precio en la ventana de listar articulos y agregue la ventana de agregar categorias

+++ tp2-grupal/Articulos.cs
@@ -11,7 +11,7 @@ namespace tp2_grupal
         public string nombre_articulo {get; set; }
         public string codigo_articulo { get; set; }
         public int Id_articulo { get; set; }
-        public float precio_articulo { get; set; }
+        public decimal precio_articulo { get; set; }
         public string descripcion_articulo { get; set; }
         public Categoria categoria_articulo { get; set; }
         public Marca marca_articulo { get; set; }
